Add basic program and README
This is the first commit of wthr, a simple command-line program that
outputs weather information. It currently only supports searching by
city.

wthr makes use of OpenWeatherMap to get weather data. I chose to use
that because it seems like a nice service. The code for fetching the
weather is based off of http://howistart.org/posts/go/1.

The code does not have any tests yet, but that is something I am going
to work on in the future.

This adds a basic README which will continue to be updated with more
information surrounding using wthr and developing it.