Add Travis CI config
This adds the config file for testing the app on CI. This will make
seeing that the tests are passing and the code is well formatted, as
well as other checks, easier because it is automated.

I chose to go with Travis CI because I like them as a company and think
the interface is nice.