Add weather condition output
This adds the current weather condition as part of the output of the
current weather. It currently only supports outputting one condition -
the first one of many that could be returned. The support for combining
conditions could be added in the future.

This adds a new condition type that represents what the weather
condition returned from OpenWeatherMap looks like.

How output now looks:

    $ wthr --city New York
    Current weather in New York is: 51°F, light rain